Change users/movies_spec.rb to movies/index_spec.rb
